requirements.txt
tensorflow
numpy
pillow
emnist
matplotlib
